Add comment version of node2vec.py
comment is for understanding the return parameter of alias_setup and simulate_walks. Has to say that, the sorted(G.neighbors(x)) is a must.
We should pay extra attention to the return tuple of alias_setup, one of which is the alias node index in sorted(G.neighbors(x)) and the other is the transition probability list.